Dpkg::OpenPGP::Backend::Sequoia: Add sqv support for verification
This adds support for the verification-only command from Sequoia-PGP,
which is the one currently used by apt to perform OpenPGP verification,
so that we can get back for example to source package verification with
dpkg-source in minimal installations.

This is possible due to the new sqv 1.3.0 API changes, which makes it
somewhat more compatible with «sq verify», in addition to supporting
inline-verify style signatures, which is required for the artifacts
that we need to handle.