libdpkg: Cache root user/group to UID/GID mapping to speed up unpacking
On glibc systems, each time we need to resolve from a user/group name to
their uid/gid, which happens for most unpacked files, we end up opening,
parsing and closing the passwd and group databases. If we cache the UID
mapping for the "root" user and group, even though on some systems the
root group is called either "wheel" (on macOS) or "system" (on AIX), we
should skip a bunch of repeated and unnecessary work for every unpack.