dpkg-deb: Add support for filesystem metadata checks
This adds checks for unusual and/or unexpected ownership of the root
directory for the to be built package.

This will help catch invocations where the resulting filesystem metadata
would be wrong and potentially a security issue, as the build has not
been made with a root-equivalent runner, such as fakeroot, or preferably
by passing the --root-owner-group option. This is part of the effort to
default to rootless builds, and to try to help external packaging efforts
that might not be using Debian tooling update their build scaffolding.