Dpkg::BuildDriver::DebianRules: Change default R³ value to «no»
The transition to rootless builds in Debian has been going on for
years now, with increasing steady adoption, although there is still
a long tail of packages that need to be adapted.

The main reason the default was never changed before, was to avoid
potential mass build breakage, and potential silent breakage in
successful builds. The former is being handled in Debian through a
normal transition (which only affects a subset of the remaining
non-transitioned packages), and the latter is helped by the new
dpkg-deb warning that will trigger when the root directory is owned
by a non-root user/group, which should catch also third party builds.

Changing the default is also the better option, and friendlier for the
packaging ecosystem to avoid requiring to set the R³ field to «no» for
every package. It also removes the need for fakeroot, while a very
helpful tool, it introduces a layer of run-time interposition that can
affect builds and introduce regressions of its own, and needs to be
kept in sync with the system libc.

Proposed-by: Niels Thykier <niels@thykier.net>