Dpkg::Vendor: Remove backwards compatibility module and origin name mangling
The naming for origin files and vendor Perl modules was rationalized
during the 1.21.x release cycle, and was slated for removal in the
1.22.x release cycle (the current one).

The code has been emitting warnings when using the affected names, and
the documentation has mentioned this as well.

It seems safe to remove the backwards compatibility code.