src/sh: Move setup_colors() contents into dpkg-error.sh script scope
The function was not written to handle being called multiple times to
perform different setups based on different inputs. Any script sourcing
this file was required to call this function as it setups variables
used by the various functions, which is unnecessary boilerplate that
we can get rid of.

Preserve an empty setup_colors() function for backwards compatibility
with code calling it.