Dpkg::OpenPGP::Backend::GnuPG: Handle and warn on concatenated ASCII Armor blocks
There are many "keyrings" in Debian source packages that contain
concatenated ASCII Armor blocks, which are problematic for interop,
but not supporting them now is a clear regression.

For now adapt the code to cope with those, and emit a warning and
hint on how that can be fixed. We cannot use SOP in the hint example,
precisely due to the interop concerns, we use «sq» because that's
the easiest and more ergonomic way to merge the certificates.