scripts: Change and inject default value for Section and Priority fields
When generating a DEBIAN/control, .dsc, .changes or a Sources repository
metaindex, set the default values for the Section and Priority fields to
«unknown» and «optional» respectively.

For the Priority field, this means that the package maintainer does not
longer need to add that field if it uses the default value, and only
need to add it to override it.

For the Section field, currently this only unifies its value when unset
so that it always is «unknown» instead of sometimes being that and
others being «-». In the future this makes way to be able to add some
heuristics to auto-generate some more useful defaults.

Proposed-by: Niels Thykier <niels@thykier.net> (for Priority)
Proposed-by: Guillem Jover <guillem@debian.org> (for Section)