Dpkg::OpenPGP::Backend::SOP: Use the parent class armor/dearmor implementations
We cannot use the SOP armor/dearmor interfaces, because concatenated
ASCII Armor is not a well supported construct, which not all SOP
implementations support. But we still need to handle this given the
data we are managing. Remove these implementations for now and use
the generic parent implementations. Once we can guarantee that our
data has been sanitized, then we could switch back to use pure SOP
interfaces.