Cast char to unsigned char for ctype functions
Whether char is signed or unsigned is platform specific. The ctype
functions take an int, but passing a negative number other than EOF
is undefined. Thus we cannot simply pass an element from a «char *»,
we need to cast it to «unsigned char» to make sure we are only ever
passing valid values.

Warned-by: gcc on NetBSD
Changelog: internal