Dpkg::Vendor::Debian: Remove merged-/usr taint logic
Now that Debian has done more or less a "proper" /usr-move, where the
symlinks are actually provided by a package, and all packages install
on the canonical location, where the packaging system should not incur
in aliasing, the warning does no longer serve much of a purpose, as
any derivative will be picking up the same changes.

Aliasing from external packages can still occur, but banning that will
be implemented via filesystem metadata tracking, once those changes can
be deployed and are not guaranteed to fail due to file type conflicts
due to the previous /usr-merge approach.

Aliasing from the filesystem is still problematic, but that should now
be either a UI issue or a reliability issue for other tools to deal
with.