scripts/mk: Use flavor to avoid re-caching empty variables
We were using the or function for the variable caches, but if the
variable is empty we might end up re-evaluating the contents, which
means we can end up executing the same command many times.

This happened to work fine with GNU make 4.3 and earlier, but stopped
working correctly with GNU make 4.4, which incurs a significant slow
down due to excess execution of for example dpkg-buildflags. This change
alone does not restore the previous number of spawned processes, but
getting back to that, implies a semantics change that can break existing
code.

Ref: #1092051