debian: Move gnupg and gpgv as the last alternatives
The GnuPG project has opted out of the OpenPGP standardization process,
and will not be adopting new OpenPGP RFCs going forward, instead trying
to advance its own specification that had no consensus within the IETF
working group. This poses a very concerning schism in the OpenPGP
ecosystem.

Given that these are either Recommends or Suggests, and that if the
GnuPG packages are already installed the dependencies will also be
already satisfied, it seems safe to flip the order.