Dpkg::BuildDriver::DebianRules: Fix R³ namespace check to only refuse unknown dpkg keywords
This was originally a problem in dpkg-buildpackage, which had been
fixed with commit dcfbe8b626cd6c3551afe7cd46500bdd38d4c200. When
refactoring the build driver, which lived in a parallel branch already
at the time of that fix, it was missed and not ported when that
branch got merged into main.

Fixes: commit 197fa86fc71a2e46a5c70f99a594bf13ffb520f9
Closes: #1092653