Dpkg::OpenPGP::Backend: Support disabling cmd and cmdv via 'none' value
For testing purposes, to be able to exercise different code paths,
it is desirable to be able to disable either cmd or cmdv with a value
that we can be certain will never match on the filesystem. Instead of
hoping that «none» will not exist on disk, handle it explicitly in the
detection code.