Dpkg::OpenPGP::Backend::GnuPG: Do not require cmdstore for backend command
We do not require the cmdstore to be present whenever we do many of the
cmd operations, and when we require it we have a dedicated function to
check whether we can use secret material.

This means we can now use the gpg backend command even when gpg-agent is
not installed.