s-s-d: Move <compat.h> and <dpkg/macros.h> after OS and header setup macros
We are setting several macros to change the behavior of headers we are
including, and because we might end up including multiple of them
depending on the compatibility code enabled or for types from the
<dpkg/macros.h> header, these might end up not taking effect, later
on due to header re-inclusion protection.

So we need to define those as early as possible, which means just after
<config.h> and before anything else.

This fixes a build failure on NetBSD 10.1, where the _KMEMUSER was not
taking effect and the <proc.h> was missing definitions for the vaddr_t
type.

Changelog: porting