s-s-d: Reformat license/copyright comment header
Make the header easier to parse, by placing author entries per line,
with intended comments for the license grants, and the changes.

While keeping the changelog here does not follow the project style,
we leave this in for now, until the general license situation of the
file gets clarified, where ideally it would get a proper license, then
the changelog can simply be removed.

Changelog: internal