Dpkg::OpenPGP::Backend::Sequoia: Do not run sq/sqv to verify with no keyrings
Both sq and sqv support expect a keyring to be able to verify the
signatures, and produce rather confusing and alarming diagnostics,
which by default are emitted as warnings.

In case we have no keyrings to pass (because they are not present on
disk for example), skip the commands invocation, and return an error
to the effect that no valid signatures could be found. In the future
we should instead print the keyrings that we are using and the ones we
are skipping, and print a specific error message for this condition,
to make the error condition more clear, but for now this is the minimal
change that gives a less confusing UI.

Closes: #1106148
Stable-Candidate: 1.22.x
(cherry picked from commit 6efc85c783b7a1b3aa4f18bb3a9be5faa2989dea)