Dpkg::OpenPGP::Backend::Sequoia: Run sq in stateless mode for verification
For verification and for every other OpenPGP command we always run in
stateless mode, by for example creating a temporary home directory to
avoid the tools (such as GnuPG), to use any implicit keyrings that we
have not explicitly passed.

But for sq we have not been doing that, which changes this specific
backend command semantics compared to the others.

Explicitly pass «--home=none» to enable stateless mode.

Suggested-by: Neal H. Walfield <neal@sequoia-pgp.org>
Stable-Candidate: 1.22.x