dpkg-source: Add support for --signer-certs option
This option makes it possible for users to specify a local trusted
keyring when they do not have access to the current vendor keyring.
And avoids having to rely on an implementation specific trusted keyring
with custom and unportable file formats such as the GnuPG trustedkeys
keyring in KeyBox format.

Closes: #1110172