Add vendor specific support for fuzzy source vs version nativeness
There's been an ongoing push to distort the native source package
concept, where this is a source of both accidental source package
construction, where even warnings have been shown to not be enough.

It has been argued this is a requirement for mainly two broad cases,
one is for native metapackages tracking non-native packages, where
the desire is to use the same versioning scheme (but with no version
remapping), and for either CI or VCS based workflows, where some of
the people involved either consider none of the proposed alternative
solutions acceptable, or consider that source packages are obsolete
and should not even be used anymore anyway, and undermining them is
clearly an acceptable price that others should pay.

This sets us back into making the packaging and its concepts more
difficult to grasp and handle, both for existing maintainers and
newcomers, and also for tool maintainers.

We need to add the notion of fuzzy nativeness, where a native source
package does not have a coherent native source version, so that vendors
that are forcing this sloppiness through, can have it without affecting
the default non-vendor case. Unfortunately this affects Debian and
derivatives. Which means using a method like Dpkg::Version->is_native()
has unreliable and non-portable semantics, so we deprecate that now
as well, and will be removed in the future.

Closes: #737634