libdpkg: Do not segfault when adding triggers in no-act mode
When we activate a trigger by calling the dpkg-trigger command with
deferred triggers in place in no-act mode, the code was not opening
the deferred triggers file for writing, but continued with the
processing, and then the subsequent attempt to write into it caused
a segmentation fault.

The root problem is that in no-act mode, we should not even attempt
to write anything. So we make the print into the file conditional on
whether the file handle is not-NULL.

[guillem@debian.org:
 - Adapt test suite for missing DPKG_DIR_ADMIN, DPKG_COMMON_OPTIONS and
   DPKG_CHECK_DIFF m4 macros, and test relocation in source tree. ]

Closes: #1108192
Stable-Candidate: 1.20.x 1.21.x 1.22.x
(cherry picked from commit b7389d6fef8d27e1cabc76b3035db0d575d26b31)