build: Automatically set test parallelism from make parallelism
Instead of having to specify the parallelism twice, with «make -j» and
then also with TEST_PARALLEL, infer the latter based on the make(1)
parallel jobs running.

On BSD make we can use the internal .MAKE.JOBS variable, and on GNU make
we can get the -j argument from the MAKEFLAGS variable which gets
exported to the make child processes.