Dpkg::Shlibs: Remove DEB_TARGET_ARCH handling from setup_library_paths()
This code was always wrong, but was added at the time to workaround
missing cross-compilation support in gcc packaging in ancient Debian
versions, that had not been updated to use the expected variables.

This has not been an issue for a long time now, and it is actually
posing problems with cross-building for different operating systems.

Reported-by: Helmut Grohne <helmutg@debian.org>