dpkg: Improve maintscript abort-* execution diagnostics
In case we end up generating error output during abort-* maintscript
invocations, that do not result in the maintscript failing, this can
become rather confusing when trying to diagnose issues.

Given that these actions will always end up with the package action
failing, and should thus not be a common occurrence, we can be more
verbose in trying to state what is going on.

Otherwise it might seem that a different maintscript action was the
cause of the error.

Closes: #1109564