dpkg: Use maintscript name instead of untranslated description in error messages
Do not use a descriptive name for the maintainer scripts, which was
currently not marked for translation and seems more confusing. The only
case where this might have mattered was to distinguish maintscripts for
the disappearing case, but it does not look necessary as we are already
printing a notice before running them.

Instead use the actual maintscript name, which is what people are going
to be more familiar with, and requires no mental mapping to the actual
filename in the «.deb» or within the source package.