dpkg: Rename maintscript execution functions
Prefix maintscript execution functions with «maintscript_run_», and
then use «old» and «new» to refer to the installed and the currently
being unpacked instances, to get more consistency, and to avoid
confusing the previously named «maintscript_new» with some kind of
constructor/allocator. Use _old_or_new instead of _fallback to make
it more clear what the actual behavior is.

Changelog: internal