libdpkg: Add support for debug_at() to print debug messages at a function
Rename the debug function to debug_print, so that we can add a new
debug_at macro that prints a debug message with the current function
name as context. This decouples the function name from the debug messages,
so that renaming the function does not require touching all such messages,
and means they will not get out of sync. Using «%s():» makes it more
clear this is the function the debug message is being emitted from.