Dpkg::Source::Package: Remove debian/rules fixer at extract time
This is already handled at build time if necessary by the relevant build
driver (Dpkg::BuildDriver::DebianRules). This means we do not emit
spurious warnings for source packages using a build driver that does not
require a debian/rules file.

The alternative to call the build driver pre-check code here, seems
architecturally wrong, because we are not building anything yet. This is
also a check for a bug that needs fixing, so reporting it later should
be fine as well. The only scenarios that would be affected are explicit
calls to debian/rules before dpkg-buildpackage has been called (which
we are trying to make an unsupported interface anyway), or iff people
want to use a read-only source tree with an overlayfs on top to avoid
modifications to files. Both of which still go back to supporting a
buggy package that needs to be fixed anyway.

Closes: #1078764