Dpkg::OpenPGP: Do not run verify with no keyrings
The verification method makes little sense without keyrings as then we'd
not be able to verify the signatures.

We implemented this specifically for Sequoia, and not for GnuPG as that
backend does not error out if it does not get a keyring, but the SOP
backend has the same problem, so let's unify the behavior for all
backends (in case users load them explicitly) and their parents which is
the expected interface for those methods.

Closes: #1111617
Stable-Candidate: 1.22.x