Dpkg::IPC: Deprecate nocheck option and rename it to no_check
This unifies the option name with the rest of similar negated option
names.

The nocheck option will emit a deprecation warnings. But for backwards
compatibility, the nocheck option can still be passed alongside no_check,
which will disable the deprecation warning, as the new option will be
used instead (now and in the future when the old option gets obsoleted).