Dpkg::Compression: Move global regexes into compression_get_file_extension_regex
One of the helper variables to construct the actual regular expression
has an extremely generic name, which is not appropriate for a file
global variable.

The other variable used to be exported, but that has not been the case
for some time, so we can move both of them as local variables inside
compression_get_file_extension_regex(), with shorter but more
descriptive names.