arch: Remove kopensolaris-any support
This port got started but it does not look like it went very far.
The OS Dyson project, which supposedly still had plans to continue
this port, has not replied to previous queries on its status, and
the glibc repositories for this port have not seen any changes for
a long time.

If the port ever comes back to life, it can always be added back.

Ref: commit 602261f1f3e3143b0b668d3ae185fb600b4ed18c
Ref: https://github.com/dtbartle/glibc-opensolaris (2009-04)
Ref: https://github.com/ip1981/kopensolaris-glibc (2015-01)