dpkg-split: Obsolete --msdos option which no longer does anything
This is a relic from the past, where distributions had to possibly
support installation from MS-DOS systems, or transport binary packages
in FAT formatted media.

This has not been the case for a very long time. And front-ends like
apt do not even support handling MSDOS-Filename in repositories, and
VFAT support is wide-spread nowadays.