Dpkg::Vendor: Add branch hardening flags to LDFLAGS
Some branch hardening features might need to also touch things up at
link time, at least to surface which features were enabled in the
module in the ELF notes.

This becomes an issue with GCC 15 which enables GCS, as the linker will
issue warning if trying to link objects with mixed GCS flags. In
essence, this would fail:

  cc $(CFLAGS) -c foo.c -o foo.o
  # Produces a lib without GCS bit
  cc $(LDFLAGS) foo.o -o libfoo.so
  …
  # Attempts to link testexec with GCS enabled
  cc $(CFLAGS) $(LDFLAGS) -lfoo -o testexec test.c

Output:

  foo.so: warning: GCS is required by -z gcs, but this shared library
  lacks the necessary property note. The dynamic loader might not enable
  GCS or refuse to load the program unless all the shared library
  dependencies have the GCS marking.

[guillem@debian.org: Formatting changes to commit message. ]

Closes: #1115292
Fixes: https://bugs.launchpad.net/bugs/2122602
Signed-off-by: Guillem Jover <guillem@debian.org>