dselect: Rewrite all methods install scripts from shell to Perl
Most of these method scripts were already in Perl, but because they were
embedded in shell scripts they could not be checked by the authorchecks,
for syntax and style issues.

Besides the simple rewrites of the shell parts into Perl, we also need
to fix the quotes so that perlcritic does not fail now, which could not
easily be fixed before as the Perl code itself was within quotes. And
update the copyright statements.

Changelog: internal
Warned-by: perlcritic
Fixes: ValuesAndExpressions::ProhibitInterpolationOfLiterals