man: Improve supported since version information
Move supported since version information into its own paragraph at the
end of the section describing the relevant item, because while it is
important information to have available it is not usually necessary. This
also makes it stand out more clearly on its own.

Unify by moving the wording into a couple of patterns, so that it is easier
to search for. These patterns are:

  Supported since (dpkg|version) <version>.
  Support for <feature> since dpkg <version>.

Not all instances have been split for now. In particular the cases where
it seemed more clear to keep the supported since information close to the
feature being described.