dpkg: Mark reinstreq during unpack as late as possible, not before prerm
Previously the package would be reinstreq while we deal with conflictors'
prerms and deconfiguration, but that's unnecessary.

During cleanup, do not pointlessly clear the reinstreq flag on postinst
"abort-upgrade", "abort-remove in-favour" and "abort-remove".
cu_prermupgrade() is only be called via a push_cleanup() in unpack.c
which is only executed if the package is at least halfconfigured so
reinstreq must be clear to start with.

[guillem@debian.org:
 - Rebase patch against current codebase.
 - Reword commit message. ]

Signed-off-by: Guillem Jover <guillem@debian.org>