Dpkg::Control::FieldsCore: Deprecate SC field export rules in binary stanza
Exporting fields into .dsc and .changes files declared in the
binary stanza of a source package template control file, makes little
sense, because there if there are multiple binary stanzas the resulting
value will depend on how the file is evaluated. If there's ever a need
for this, then proper support needs to be added to dpkg, so that any
value can be coalesced to generate a single output value.

Prompted-by: Richard Hansen <rhansen@rhansen.org>
Ref: https://bugs.debian.org/1117566