Dpkg::BuildInfo: Allow TZ, TZDIR and DATEMSK variables
Allow timezone and date related environment variables. While the TZ
environment variable could be considered to leak local user information,
it only tends to be used to override the system default, and we are
already leaking it indirectly via the deb-buildinfo(5) Build-Date field
for all builds, and for maintainer builds this is also leaked via the
trailer line in the changelog entry.