libcompat: Add countof() support
This is currently defined in the upcoming C2y (C29) standard draft, as
a macro using the _Countof built-in, that counts the number of elements
from an array.

We are currently using array_count(), but using the new name will mean
that we can already use the support from compilers that have implemented
it, and can eventually rely exclusively on the language support, instead
of on our open-coded fallback implementation.