Standardize error/warning messages on "cannot ..."
Unify all error/warning messages on a single formula, so that it is
easier to read, reduces translation work, and it is easier to write
new messages w/o having to think too much. Use "cannot ..." instead of:

   "couldn't ..."
   "could not ..."
   "can't ..."
   "can not ..."
   "failed to ..."
   "unable to ..."
   "error <doing something ...>"
   "... failed"

We do not change errors from external processes, where we have no idea
what went wrong, and leave them as "<process> failed <with something>".