libdpkg: Rework compression error messages
Change the messages to explain what action we cannot do, instead of
referring to an ambiguous "internal <something> error". When possible
mention whether we are writing or reading compressed or decompressed
data, and refer to these inputs and outputs as "<compression> stream".
For generic errors, use "<compressor> failure" instead of "error", as
we are already printing that as a prefix.

Add a TRANSLATORS note to clarify string that might be otherwise hard to
translate w/o context.