Stop using length limited format strings (%.255s/%.250s/%.100s/%.50s)
These have historically been used for strings coming from external
sources, where they might exceed a certain length, which could cause
truncation of the overall message in emergency situations, where we have
reserved 4 KiB for that scenario.

The problem is that this will always truncate argument strings that
currently exceed those limits, which seems worse, as the end of the
string tends to be the most significant part for example in pathnames.
For example in Debian there are already packages with pathnames
exceeding 255 characters (up to 305 characters).

Given that this is should only be an issue with error conditions under
high memory pressure, where we will most certainly end up aborting due
to the lack of memory, stop trying to limit those strings in the common
case.

Changelog: internal