dpkg-source: Add a new profile:v1 property in Package-List field
The profile property has several problems. It uses the plus character
which is valid in package names, which can then be parsed incorrectly.
And the separator characters are non-obvious mappings.

We cannot replace value format for this property as that would break
any existing user. Instead we add a new version 1 for this property,
and use a more obvious and package name safe set of characters.

New characters suggested by Simon McVittie as the more obvious
candidates to use.

Ref: https://lists.debian.org/debian-devel/2025/11/msg00333.html