man: Improve architecture documentation
Try to unify the terminology used to refer to the various concepts.
Add some notes about specific values not applying in some contexts.

Describe first what is a Debian architecture, then specify the allowed
values and their use.

Update some example architectures with names that might be easier to
come across as being currently in use.

Prompted-by: Helmut Grohne <helmut@subdivi.de>