Dpkg::BuildProfiles: Add workaround for callers passing invalid formulas
The dh-exec tool, makes use of this module, but because the parser was
not strict, it has accepted strings that were not valid build profile
formulas.

This is currently making packages fail to build from source. As a
temporary workaround check whether the passed string is not valid,
and try to fix it up by surrounding it with angle brackets.

Diagnosed-by: Chris Hofstaedtler <zeha@debian.org>