Dpkg::Source::Package::V1: Make debian/rules executable on extract if present
The 1.0 source format uses a plain patch to transport the debian/
directory, where by default we have been creating that patch with a diff
tool that is not able to represent executable bits, necessary for the
debian/rules file.

This means that now that we removed the fixed from the common
Dpkg::Source::Package module, we always get a warning from
dpkg-buildpackage when doing a build, which is not correct, as there is
nothing the maintainer can do about it, contrary to newer source
formats.

Add back the debian/rules fixer, but make it specific to the 1.0 source
format, and make it not warn at all if it is missing or not already
executable. If the debian/rules file is either missing or not executable
by the time we are doing the build, then we can safely warn on either
condition.

Fixes: commit 91f68ddf6537e28bf2283a983d64f42494652b96
Closes: #1123630